Попытка написать заготовку для "высокоуровневого сервиса для добавления комментариев". Возможно, его надо будет сделать более интеллектуальным (например, добавлять комментарий не по идентификатору книги, а, скажем, по совпадению подстроки в названии).

+++ dz08/src/main/java/ru/otus/spring08/dao/BookDaoJpa.java
@@ -31,7 +31,7 @@ public class BookDaoJpa implements IBookDao {
     }
 
     @Override
-    public Book getById(int id) {
+    public Book getById(Long id) {
         return em.find(Book.class, id);
     }
